requirements.txt
feedparser==6.0.11
requests==2.32.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
beautifulsoup4==4.12.3
lxml==5.3.0
pypdf==4.3.1
